Item Apply 로직 설계
1. Item 실행 순서 정리
- ItemPrefab : ItemPickup.OnTrigger -> ItemData.Apply -> PlayerItemChecker -> 아이템 효과 적용 to Player
- 해당 연결과정 추가 정의 필요

2. CoinMagnet 아이템 별개 로직 정리
- CoinMagnet 효과 적용을 위한 자식 오브젝트(CoinController) 설계
- 해당 오브젝트 Collider 의 Radius 범위 증가 로직 필요

+++ Assets/Scripts/Enemy/BaseEnemy.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 /// </summary>
 public interface IDamageable
 {
-	void TakeDamage(int amount);
+	void TakeDamage(float amount);
 }
 
 /// <summary>
@@ -28,7 +28,7 @@ public abstract class BaseEnemy : MonoBehaviour, IDamageable, IDamageDealer, IPo
 	{
 		enemyCurHealth = enemyMaxHealth;
 	}
-	public virtual void TakeDamage(int amount)
+	public virtual void TakeDamage(float amount)
 	{
 
 	}
